requirements.txt
Jinja2~=3.1.2
requests~=2.32.3
GitPython~=3.1.44
PyGithub~=2.5.0
